+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Add support for mentioning users in comments. Typing `@` (or pressing `ctrl+@`) while writing a comment on
+Jira Cloud (API v3) opens a user search; selecting a user inserts an `@[Name](accountId)` token that is
+submitted as an ADF `mention` node. Resolves [#125](https://github.com/whyisdifficult/jiratui/issues/125). By [@mrsnakeoil](https://github.com/mrsnakeoil)
+
 ## [1.14.0] 2026-09-13
 
 ### Added